+++ src/main/java/com/vadmack/mongodbtest/controller/ProjectController.java
@@ -1,0 +1,48 @@
+package com.vadmack.mongodbtest.controller;
+
+import com.vadmack.mongodbtest.dto.ProjectDto;
+import com.vadmack.mongodbtest.entity.Project;
+import com.vadmack.mongodbtest.service.ProjectService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/projects")
+public class ProjectController {
+
+    @Autowired
+    private ProjectService service;
+
+    @GetMapping
+    public ResponseEntity<List<Project>> findAll() {
+        return ResponseEntity.ok(service.findAll());
+    }
+
+    @GetMapping(value = "/{id}")
+    public ResponseEntity<Project> findById(@PathVariable(value = "id") Long id) {
+        return ResponseEntity.ok(service.findById(id));
+    }
+
+    @PostMapping
+    public ResponseEntity create(@RequestBody ProjectDto projectDto) {
+        service.create(projectDto);
+        return new ResponseEntity(HttpStatus.CREATED);
+    }
+
+    @PutMapping(value = "/{id}")
+    public ResponseEntity update(@PathVariable(value = "id") Long id,
+                                 @RequestBody ProjectDto projectDto) {
+        service.update(id, projectDto);
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
+    }
+
+    @DeleteMapping(value = "/{id}")
+    public ResponseEntity delete(@PathVariable(value = "id") Long id) {
+        service.delete(id);
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
+    }
+}
